add profile components
