+++ ngapp/src/app/profile/profile/profile.component.html
@@ -1,62 +1,14 @@
-<div class="profileContainer" *ngIf="auth.getUserDetails()">
-    <!-- Profile options -->
-    <div class="formTitle">{{ts.l.profile}}</div><br>
-    <div class="formBody">  
-      <div class="profileFormContainer">
-        <!-- error message container -->
-        <div class="usernameErrorText"> {{errorMessage}} </div>
-        <!-- username details -->
-        <div class="form-group">
-          <div class="formLabel">{{ts.l.username}} :</div>
-          <div class="userInfo">
-            <i>{{auth.getUserDetails().username}}</i>
-            <i class="fas fa-edit headerIconBtn" (click)="openUpdateUsernameDialog()"></i>
-          </div>
-        </div>
-        <!-- classroom details if STUDENT -->
-        <div class="form-group" *ngIf="classroom && auth.getUserDetails().role === 'STUDENT'">
-          <div class="formLabel">{{ts.l.classroom}} :</div>
-          <div class="userInfo"><i>{{classroom.title}}</i></div>
-          <button class="btn leaveClassroomButton" (click)="leaveClassroom()">{{ts.l.leave}}</button>
-        </div>
-        <!-- join classroom option STUDENT -->
-        <div class="form-group-vertical" *ngIf="!classroom && auth.getUserDetails().role === 'STUDENT'">
-          <div class="form-group-section">
-            <div class="formLabel">{{ts.l.classroom_code}} :</div>
-            <input type="text" class="form-control inputField" placeholder="{{ts.l.enter_code}}" name="classroomcode" [formControl]="codeInput">
-          </div>
-          <div class="classroomMessages">
-            <div *ngIf="classroomCodeOutput" class="notFoundText classroomMessage">{{classroomCodeOutput}}</div>
-            <div *ngIf="foundClassroom" class="greenLink classroomMessage" (click)="joinClassroom()">{{ts.l.join_classroom}} '<b>{{foundClassroom.title}}</b>'?</div>
-          </div>
-        </div>
-        <!-- change language -->
-        <div class="form-group">
-          <div class="formLabel">{{ts.l.language}} :</div>
-          <div *ngIf="ts.inEnglish()" class="languageButtonContainer">
-            <div class="languageButton languageButtonSelected">{{ts.l.english}}</div>
-            <div class="languageButton" (click)="ts.setLanguage('ga')">{{ts.l.irish}}</div>
-          </div>
-          <div *ngIf="ts.inIrish()" class="languageButtonContainer">
-            <div class="languageButton" (click)="ts.setLanguage('en')">{{ts.l.english}}</div>
-            <div class="languageButton languageButtonSelected">{{ts.l.irish}}</div>
-          </div>
-        </div>
-        <div class="accountButtonContainer">
-          <!-- delete account -->
-          <div>
-            <button class="btn deleteAccountButton" (click)="openDeleteDialog()"> {{ts.l.delete_my_account}} </button>
-          </div>
-          <!-- Change password -->
-          <div>
-            <button class="btn passwordButton" (click)="openUpdatePasswordDialog()"> {{ts.l.change_password}} </button>
-          </div>
-        </div>
-        <!-- sign out and profile information -->
-        <div class="profileBtnContainer">
-          <button routerLink="/register-profile" class="btn logoutBtn">{{ts.l.my_info}}</button>
-          <button routerLink="/landing" class="btn logoutBtn" (click)="logout()">{{ts.l.sign_out}} <i class="fas fa-sign-out-alt"></i></button>
-        </div>
-      </div>
-    </div>
+<div class="profileContainer">
+  <div class="sideMenu">
+    <div (click)="displayComponent = 'accountSettings'">Account Settings</div>
+    <div (click)="displayComponent = 'languageInfo'">My Language Info</div>
+    <div (click)="displayComponent = 'learningGoals'">Learning Goals</div>
+    <div>Sign Out</div>
+  </div>
+  <div class="body" [ngSwitch]="displayComponent">
+    <app-account-settings *ngSwitchDefault>account settings</app-account-settings>
+    <app-register-profile *ngSwitchCase="'languageInfo'">language info</app-register-profile>
+    <div *ngSwitchCase="'learningGoals'">Coming Soon!</div>
+  </div>
+
 </div>
